esh-autosuggest--prefix: Fix input-start
In Emacs 30+, `eshell-bol` is merely an alias for `beginning-of-line`,
not returning the point any longer.  This result in `input-start` always
returning nil instead of the correct position of the point.  Further,
the implementation misbehaves when there is no previous prompt at all,
hanging Emacs.  There are built in ways to query the previous and next
prompt in order to get a sensible starting position, so use these
instead.